requirements.txt
channels==2.4.0
channels-redis==2.4.2
Pillow==6.2.1
django-sslserver
selenium
requests
